+++ libft/ft_memcpy.c
@@ -6,17 +6,17 @@
 /*   By: selbouka <selbouka@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/23 11:10:31 by selbouka          #+#    #+#             */
-/*   Updated: 2024/11/15 16:32:30 by selbouka         ###   ########.fr       */
+/*   Updated: 2025/10/31 22:28:31 by selbouka         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memcpy(void *dest, const void *src, size_t n)
+void	*ft_memcpy(void *dest, const void *src, int n)
 {
 	char	*pd;
 	char	*ps;
-	size_t	i;
+	int	i;
 
 	i = 0;
 	pd = (char *)dest;
