+++ libft/ft_strcpy.c
@@ -6,20 +6,19 @@
 /*   By: selbouka <selbouka@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/18 18:40:39 by selbouka          #+#    #+#             */
-/*   Updated: 2025/08/18 18:41:35 by selbouka         ###   ########.fr       */
+/*   Updated: 2025/10/31 22:28:31 by selbouka         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strcpy(char *dst, const char *src)
+int	ft_strcpy(char *dst, const char *src)
 {
-	size_t	i;
-	size_t	len;
+	int	i;
+	int	len;
 
 	len = ft_strlen(src);
 	i = 0;
-
 	while (src[i] != '\0')
 	{
 		dst[i] = src[i];
